# Conflicts:
#	survey-frontend/src/app/api.service.ts
#	survey-frontend/src/app/questionlist/questionlist.component.ts

+++ survey-frontend/src/app/survey/survey.component.ts
@@ -7,7 +7,7 @@ import { HttpModule, Http, Response } from '@angular/http';
 import { HttpClientModule } from '@angular/common/http';
 import { Observable } from '../../../node_modules/rxjs';
 import { ApiService } from '../api.service';
-import {DataService} from '../data.service';
+import { DataService } from '../data.service';
 @Component({
   selector: 'app-survey',
   templateUrl: './survey.component.html',
@@ -19,73 +19,20 @@ export class SurveyComponent implements OnInit {
   message: string;
   srname: String;
   @Output() messageEvent = new EventEmitter();
-  /*surveyName: String = 'hello';
-  surveyMessage: String = null;
-  surveyDeclaration: string = null;
-  surveyStatus: string = null;
-  startDate: string = null;
-  expiryDate: string = null;
-  createdBy: string = null;
-  createdOn: string = null;
-  questions: string[] = null;
-  survey: Survey = {
-  surveyName: this.surveyName,
-  surveyMessage: this.surveyMessage,
-  surveyDeclaration: this.surveyDeclaration,
-  surveyStatus: this.surveyStatus,
-  startDate: this.startDate,
-  expiryDate: this.expiryDate,
-  createdBy: this.createdBy,
-  createdOn: this.createdOn,*/
-  /* questions: {
-      type: "array",
-      }, */
-
-  /*questions: this.questions
-  };*/
+  updSur: Survey = {'surveyName': '',
+'surveyMessage': '',
+'surveyDeclaration': '',
+'surveyStatus': '',
+'startDate': '',
+'expiryDate': '',
+'questions': []
+};
   srv: Survey;
   result: any;
   testsur: object;
   surv: Survey;
   sur: any = {};
-  // surveys: Survey[] = [];
-  /*surveys: Survey[];*/ /* = [{'surveyName': 'Test Survey 2',
-  'surveyMessage': 'Welcome Msg',
-  'surveyDeclaration': 'Decl',
-  'surveyStatus': 'A',
-  'startDate': '03/10/2018',
-  'expiryDate': '23/10/2018',
-  'createdBy': '5bb5e0ead72cc717cc3be713',
-  'createdOn': '03/10/2018',
-  'questions':
-  [
-  '5bb49f110d9e321bcc8055b5',
-  '5bb5e1f1d72cc717cc3be714',
-  '5bb5e2afd72cc717cc3be718'
- ]},
- {'surveyName': 'Test Survey 3',
- 'surveyMessage': 'Welcome Msg3',
- 'surveyDeclaration': 'Decl3',
- 'surveyStatus': 'A',
- 'startDate': '03/10/2018',
- 'expiryDate': '23/10/2018',
- 'createdBy': '5bb5e0ead72cc717cc3be713',
- 'createdOn': '03/10/2018',
- 'questions':
- [
- '5bb49f110d9e321bcc8055b53',
- '5bb5e1f1d72cc717cc3be7143',
- '5bb5e2afd72cc717cc3be7183'
-]}];*/
-  /*stitle: String = '';
-  smessage: String = '';
-  ssdate: String = '';
-  esdate: String = '';
-  decl: String = '';
-  sts: String = '';
-  qs: String[] ;*/
-
-   surveys: Array<object> = [];
+  surveys: Array<object> = [];
 
   constructor(private apiService: ApiService,
     private _router: Router,
@@ -102,35 +49,26 @@ export class SurveyComponent implements OnInit {
     // this.surveys.push(this.sur);
 
     // this.http.get(this.readUrl + '/survey/survey').subscribe(data => { this.result = (data['_body'] | JSON);
-       // this.surveys.push(data['_body']|JSON);
+    // this.surveys.push(data['_body']|JSON);
     // console.log(data);
-  // console.log(data['_body']);
-  // });
+    // console.log(data['_body']);
+    // });
 
-   /*  this.http.get(this.readUrl + '/survey/survey').subscribe(data => {
-      this.surveys.push(data['_body']);
-      //console.log(data);
-      console.log(data['_body']);
-    }); */
+    /*  this.http.get(this.readUrl + '/survey/survey').subscribe(data => {
+       this.surveys.push(data['_body']);
+       //console.log(data);
+       console.log(data['_body']);
+     }); */
   }
 
   public getSurveys() {
     this.apiService.getSurveys().subscribe((data: Array<object>) => {
       this.surveys = data;
-      // console.log(data);
+       console.log(data);
     });
 
   }
   addsurvey(frm): Observable<Response> {
-    /*this.surveys.push({
- 'surveyName': frm.surveyName.value,
- 'surveyMessage': frm.surveyMessage.value,
- 'surveyDeclaration': frm.surveyDeclaration.value,
- 'surveyStatus': frm.surveyStatus.value,
- 'startDate': frm.startDate.value,
- 'expiryDate': frm.expiryDate.value,
- 'questions': []
-    });*/
     this.surv = {
       'surveyName': frm.surveyName.value,
       'surveyMessage': frm.surveyMessage.value,
@@ -142,7 +80,7 @@ export class SurveyComponent implements OnInit {
     };
     console.log(this.surv);
     this.http.post(this.readUrl + '/survey/survey', this.surv).subscribe(res => {
-    this.result = res;
+      this.result = res;
       console.log(res);
       this.apiService.getSurveys().subscribe((data: Array<object>) => {
         this.surveys = data;
@@ -157,12 +95,12 @@ export class SurveyComponent implements OnInit {
     // console.log(frm);
   }
   sendMessage(srvy: Survey) {
-  console.log(this.message);
-  this.dataservice.changeMessage(srvy);
-  this.dataservice.currentMessage.subscribe(message => this.srv = message);
-  console.log(this.srv);
-  // console.log(Name);
-  // this.messageEvent.emit(Name);
+    console.log(this.message);
+    this.dataservice.changeMessage(srvy);
+    this.dataservice.currentMessage.subscribe(message => this.srv = message);
+    console.log(this.srv);
+    // console.log(Name);
+    // this.messageEvent.emit(Name);
   }
   /*addSurvey(frm) {
     this.surveys = {
@@ -176,20 +114,48 @@ export class SurveyComponent implements OnInit {
          };
     return this._surveyser.save(this.surveys);
   }*/
-deletesurvey(id): Observable<Response> {
-  console.log(id);
+  deletesurvey(id): Observable<Response> {
+    if (confirm('Are you sure you want to delete this survey?')) {
+      //console.log(id);
 
-  this.http.delete(this.readUrl + '/survey/survey/' + id).subscribe(res => {
-    console.log(res);
-    this.apiService.getSurveys().subscribe((data: Array<object>) => {
-      this.surveys = data;
-       console.log(this.surveys);
-    });
-  });
-  return ;
+      this.http.delete(this.readUrl + '/survey/survey/' + id).subscribe(res => {
+        var body = res['_body'];
+        var response = JSON.parse(body);
+        alert(response.msg);
+        //console.log(response.msg);
+        this.apiService.getSurveys().subscribe((data: Array<object>) => {
+          this.surveys = data;
+          // console.log(this.surveys);
+        });
+      });
+      return;
+    }
+  }
+  upsur(survey) {
+    this.surv = survey;
+    console.log(this.surv);
+  }
 }
-upsur(survey) {
-  this.surv = survey;
-  console.log(this.surv);
+ upsur(survey) {
+  this.updSur = survey;
+   console.log(this.updSur);
+   console.log(this.updSur['startDate']);
+   console.log(this.updSur['expiryDate']);
 }
+ updateSur(survey) {
+   console.log(this.updSur['_id']);
+   console.log(this.updSur);
+   this.apiService.updateSurvey(this.updSur['_id'], this.updSur).subscribe(res => {
+     alert('Survey Succesfully Updated');
+     console.log(res);
+   });
+   this.updSur = {'surveyName': '',
+   'surveyMessage': '',
+   'surveyDeclaration': '',
+   'surveyStatus': '',
+   'startDate': '',
+   'expiryDate': '',
+   'questions': []
+   };
+ }
 }
